requirements.txt
asgiref==3.3.4
astroid==2.4.2
autopep8==1.5.7
black==21.12b0
blessings==1.7
bpython==0.21
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
curtsies==0.3.5
cwcwidth==0.1.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.10
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
greenlet==1.1.0
gunicorn==20.1.0
httpie==2.3.0
idna==2.10
isort==5.7.0
itypes==1.2.0
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
pyxdg==0.27
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
typing-extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.2.0
wrapt==1.12.1
